Refactor details activity to use Domain model
