+++ app/src/main/java/com/example/beerhive/beerdetail/BeerDetailActivity.kt
@@ -7,18 +7,23 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.example.beerhive.R
 import com.example.beerhive.databinding.ActivityBeerDetailBinding
+import com.example.beerhive.domain.Beer
 
 class BeerDetailActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityBeerDetailBinding
 
-    companion object{
+    companion object {
         const val BEER_EXTRA = "beer_extra"
     }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val binding: ActivityBeerDetailBinding = DataBindingUtil.setContentView(this, R.layout.activity_beer_detail)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_beer_detail)
+
+
         if (intent != null && intent.hasExtra(BEER_EXTRA)) {
             val beer: Beer = intent.getParcelableExtra(BEER_EXTRA)
-            beer.beerBrewingTipsTitle = getString(R.string.brewing_tips)
+            // beer.beerBrewingTipsTitle = getString(R.string.brewing_tips)
             binding.beerDetail = beer
             Glide.with(binding.beerImageView.context).load(beer.beerImageUrl).dontAnimate().fitCenter().diskCacheStrategy(
                     DiskCacheStrategy.RESOURCE)
