+++ app/src/main/java/com/example/beerhive/domain/Models.kt
@@ -1,17 +1,22 @@
-package com.example.beerhive.beerdetail
+package com.example.beerhive.domain
 
 import android.os.Parcel
 import android.os.Parcelable
 
-data class Beer(var beerName: String? = null,
+data class Beer(var beerName: String?=null,
                 var beerDescription: String? = null,
                 var beerBrewerTips: String? = null,
                 var beerImageUrl: String? = null,
-                var beerBrewingTipsTitle: String? = null) : Parcelable {
+                var beerBrewingTipsTitle: String? = null,
+                var volume: Int,
+                var volumeUnit: String? = null):Parcelable {
     constructor(parcel: Parcel) : this(
             parcel.readString(),
             parcel.readString(),
             parcel.readString(),
+            parcel.readString(),
+            parcel.readString(),
+            parcel.readInt(),
             parcel.readString())
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
@@ -19,6 +24,9 @@ data class Beer(var beerName: String? = null,
         parcel.writeString(beerDescription)
         parcel.writeString(beerBrewerTips)
         parcel.writeString(beerImageUrl)
+        parcel.writeString(beerBrewingTipsTitle)
+        parcel.writeInt(volume)
+        parcel.writeString(volumeUnit)
     }
 
     override fun describeContents(): Int {
